<short summary of the patch>
TODO: Put a short summary on the line above and replace this paragraph
with a longer explanation of this change. Complete the meta-information
with other relevant fields (see below for details). To make it easier, the
information below has been extracted from the changelog. Adjust it or drop
it.

expand-region-el (0.11.0-3) unstable; urgency=medium

  * update packaging: versions, swizzle to salsa

Gbp-Pq: Name debian-changes